Introduce build convention plugins for common project setup
Extracts shared Gradle configurations from the root `build.gradle.kts` `subprojects` block into dedicated convention plugins.

This refactoring centralizes reusable build logic for:
*   `triplea-test-conventions`: JUnit Platform, test logging, custom test output, and JaCoCo.
*   `triplea-base-project`: Java compiler options, Checkstyle, and PMD. This plugin also applies `triplea-test-conventions`.
*   `triplea-java-application`: Applies the `application` plugin and `triplea-base-project`.
*   `triplea-java-library`: Applies `java-library`, Spotless, and `triplea-base-project`.

This approach reduces boilerplate and improves modularity, reducing the need for extensive `subprojects` configuration and keeping project setup more local and explicit.
