+++ gradle/build-logic/triplea-base-project/src/main/kotlin/triplea-base-project.gradle.kts
@@ -1,0 +1,3 @@
+plugins {
+    id("triplea-test-conventions")
+}
