+++ gradle/build-logic/triplea-java-application/build.gradle.kts
@@ -3,9 +3,8 @@ plugins {
 }
 
 dependencies {
-    implementation(libs.spotless)
     implementation(project(":triplea-base-project"))
 }
 
-description = """This project creates a pre-compiled script plugin that defines a conventional library project used in the build.  
+description = """This project creates a pre-compiled script plugin that defines a conventional application project used in the build.  
 This avoids the need for `subprojects` and `allprojects`, and keeps project configuration local."""
