+++ gradle/build-logic/triplea-java-application/src/main/kotlin/triplea-java-application.gradle.kts
@@ -1,0 +1,9 @@
+/*
+  This convention defines a standard TripleA java application project.
+  It applies the `application` plugin and applies universal configuration.
+*/
+
+plugins {
+    id("application")
+    id("triplea-base-project")
+}
