+++ gradle/build-logic/triplea-java-library/src/main/kotlin/triplea-java-library.gradle.kts
@@ -6,6 +6,7 @@
 plugins {
     `java-library`
     id("com.diffplug.spotless")
+    id("triplea-base-project")
 }
 
 group = "triplea"
