+++ gradle/build-logic/triplea-test-conventions/build.gradle.kts
@@ -1,0 +1,5 @@
+plugins {
+    `kotlin-dsl`
+}
+
+description = """This project creates a pre-compiled script plugin that defines testing conventions used by all projects in the build."""
